Get rid of redundant structures and actions
Move schema parsing methods to `Generator`

Get rid of `validator` structure

Decrease number of walks through generator files

Remove receivers from certain functions
